dev-requirements.txt
-r requirements.txt

fabric3
factory_boy
pytest
pytest-django
